requirements.txt
python-telegram-bot==20.7 
black==23.12.1
isort==5.13.2
ruff==0.1.6  
